Refactor logo components and update branding
Introduced a new Logo component supporting multiple variants (Default, Icon, Wordmark) and updated usage across the app. Moved combination-mark to a dedicated logos directory. Added a new wordmark SVG and updated business info to include the wordmark path. Improved accessibility in the language switcher.
