+++ components/footer/_components/language-switcher.tsx
@@ -51,6 +51,7 @@ export const LanguageSwitcher = ({ params }: { params: Promise<{ lang: "en" | "e
           role="combobox"
           aria-expanded={open}
           className="w-auto justify-start gap-2 border-0 bg-[#f1a208] hover:bg-[#f1a208]/80 text-black cursor-pointer"
+          aria-label="Switch Language"
         >
           {languages.find((language) => language.value === value)?.flag && (
             <span className="inline-block align-middle">
